+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [v2.0.3-rc] - 2020-03-19
+### Fixed
+- Marking peer as offline properly when peer has created no snapshots
+
 ## [v2.0.2-rc] - 2020-03-18
 ### Fixed
 - Filter out offline peers for data resolving #947